Добавляет Smart view
